pyproject.toml
[project]
name = "ai-video-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "google-ai-generativelanguage>=0.6.15",
    "google-api-python-client>=2.179.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.30.0",
    "google-generativeai>=0.8.5",
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
]
